requirements.txt
newspaper4k==0.9.3.1
lxml[html_clean]
feedparser==6.0.11
beautifulsoup4==4.13.3
nltk==3.9.1
requests==2.32.3
python-dateutil==2.9.0.post0
tldextract==5.1.3
pandas==2.2.3
PyYAML==6.0.2
urllib3>=1.26.12
reppy>=0.4.14
pydantic>=2.0.0
fastapi>=0.95.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
sentence-transformers>=2.2.2
docker>=6.1.0
pytest>=7.3.1
mypy>=1.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
uvicorn>=0.17.6
gunicorn>=20.1.0
python-multipart>=0.0.5
pgvector>=0.2.0
alembic>=1.10.0
httpx>=0.23.0
tenacity>=8.2.0
python-dotenv>=1.0.0
groq>=0.18.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
langchain-huggingface>=0.0.1
psutil>=5.9.0
pytest-cov>=4.1.0
# Topic classification and entity extraction
spacy>=3.5.0
python-json-logger>=2.0.7
prometheus-client>=0.16.0
elasticsearch>=8.8.0
regex>=2022.10.31 